CommonDeclaration.h 삭제, 해당 파일 안에 있던 모든 데이터를 그 데이터와 가장 관련있는 클래스 앞으로 옮김.
