+++ MineSweeper.vcxproj
@@ -43,7 +43,6 @@
     <ClInclude Include="src\Combats\OddOrEven.h" />
     <ClInclude Include="src\Combats\RockPaperScissor.h" />
     <ClInclude Include="src\Combats\ShootTheMonster.h" />
-    <ClInclude Include="src\CommonDeclarations.h" />
     <ClInclude Include="src\MineField.h" />
     <ClInclude Include="src\resource.h" />
     <ClInclude Include="src\Stage.h" />
